restruct && Clean the endpoints
