+++ src/services/notifications.service.js
@@ -1,8 +1,8 @@
-const assert = require('assert');
-
 const insertNotification = function (db, data, callback) {
+
     // Get the documents collection
     const collection = db.collection('notifications')
+
     // Insert some documents
     collection.insertOne(data, function (err, result) {
         if (err) {
